selecting bows now work

+++ TheArcheryMerchant/models.py
@@ -8,7 +8,7 @@ class Bow(models.Model):
     bowtype = models.CharField(max_length=100, choices=types)
     drawLength = models.IntegerField(default=0)
     drawStrength = models.IntegerField(default=0)
-    price = models.IntegerField(default=0)
+    price = models.DecimalField(max_digits=5, decimal_places=2, default=0.0)
     sellers = models.CharField(max_length=100, choices=sellers)
     
     def __str__(self):
@@ -20,7 +20,7 @@ class Arrows(models.Model):
     arrowLength = models.IntegerField(default=0)
     spinage = models.IntegerField(default=0)
     brand = models.CharField(max_length=100)
-    price = models.IntegerField(default=0)
+    price = models.DecimalField(max_digits=5, decimal_places=2, default=0.0)
     sellers = models.CharField(max_length=100, choices=sellers)
     
     def __str__(self):
